+++ pkg/nodestack/nodestack.go
@@ -1,4 +1,4 @@
-package parsing
+package nodestack
 
 import (
 	"sort"
